TASK 6

+++ 0x00-hello_world/6-size.c
@@ -13,11 +13,11 @@ int main(void)
 	long int e;
 	long long int f;
 
-	printf("Size of int is : %lu \n", sizeof(a));
-	printf("Size of float is : %lu \n", sizeof(b));
-	printf("Size of double is : %lu \n", sizeof(c));
-	printf("Size of char is : %lu \n", sizeof(d));
-	printf("Size of a long int : %lu \n", sizeof(e));
-	printf("Size of a long long int: %lu \n", sizeof(f));
+	printf("Size of int is : %lu \n", (unsigned long)sizeof(a));
+	printf("Size of float is : %lu \n", (unsigned long)sizeof(b));
+	printf("Size of double is : %lu \n", (unsigned long)sizeof(c));
+	printf("Size of char is : %lu \n", (unsigned long)sizeof(d));
+	printf("Size of a long int : %lu \n", (unsigned long)sizeof(e));
+	printf("Size of a long long int: %lu \n", (unsigned long)sizeof(f));
 	return (0);
 }
